+++ resources/js/components/Background.vue
@@ -11,9 +11,11 @@
         <div v-if="type === 'image'">
             <div class="row">
                 <div class="col-12">
-                    <div class="custom-file">
-                        <label for="file" class="custom-file-label" :data-browse="__('Choose')">{{ __('File') }}</label>
-                        <input type="file" class="custom-file-input" id="file" @change="handleFile" accept="image/*" :value="image.name"/>
+                    <div class="form-group">
+                        <div class="custom-file">
+                            <label for="file" class="custom-file-label" :data-browse="__('Choose')">{{ __('File') }}</label>
+                            <input type="file" class="custom-file-input" id="file" @change="handleFile" accept="image/*" :value="image.name"/>
+                        </div>
                     </div>
                 </div>
             </div>